Move key rotation to km ownership
