+++ apps/web/src/app/auth/settings/settings.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from "@angular/core";
 
 import { PasswordCalloutComponent } from "@bitwarden/auth/angular";
 
+import { UserKeyRotationModule } from "../../key-management/key-rotation/user-key-rotation.module";
 import { SharedModule } from "../../shared";
 import { EmergencyAccessModule } from "../emergency-access";
-import { UserKeyRotationModule } from "../key-rotation/user-key-rotation.module";
 
 import { ChangePasswordComponent } from "./change-password.component";
 import { WebauthnLoginSettingsModule } from "./webauthn-login-settings";
